Rename liquity-ext folder to liquity
