+++ subgraphs/aave-v3/protocols/aave-v3/src/entities/position.ts
@@ -228,6 +228,7 @@ function closePosition(
   position: Position
 ): void {
   position.blockNumberClosed = event.block.number;
+  position.timestampClosed = event.block.timestamp;
   position.hashClosed = event.transaction.hash.toHexString();
   position.timestampClosed = event.block.timestamp;
   if (position.side == PositionSide.LENDER) {
