--- subgraphs/liquity-ext/protocols/liquity-extended/config/templates/liquity-extended.template.yaml
@@ -1,130 +1,0 @@
-specVersion: 0.0.4
-repository: https://github.com/messari/subgraphs
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum/contract
-    name: TroveManager
-    network: mainnet
-    source:
-      address: "0xA39739EF8b0231DbFA0DcdA07d7e29faAbCf4bb2"
-      abi: TroveManager
-      startBlock: 12178557
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      language: wasm/assemblyscript
-      entities: []
-      abis:
-        - name: TroveManager
-          file: ./abis/TroveManager.json
-      eventHandlers:
-        - event: TroveUpdated(indexed address,uint256,uint256,uint256,uint8)
-          handler: handleTroveUpdated
-        - event: Redemption(uint256,uint256,uint256,uint256)
-          handler: handleRedemption
-        - event: TroveLiquidated(indexed address,uint256,uint256,uint8)
-          handler: handleTroveLiquidated
-      file: ./src/mapping/TroveManager.ts
-  - kind: ethereum/contract
-    name: BorrowerOperations
-    network: mainnet
-    source:
-      address: "0x24179CD81c9e782A4096035f7eC97fB8B783e007"
-      abi: BorrowerOperations
-      startBlock: 12178582
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      language: wasm/assemblyscript
-      entities: []
-      abis:
-        - name: BorrowerOperations
-          file: ./abis/BorrowerOperations.json
-      eventHandlers:
-        - event: LUSDBorrowingFeePaid(indexed address,uint256)
-          handler: handleLUSDBorrowingFeePaid
-        - event: TroveUpdated(indexed address,uint256,uint256,uint256,uint8)
-          handler: handleTroveUpdated
-      file: ./src/mapping/BorrowerOperations.ts
-  - name: PriceFeed
-    kind: ethereum/contract
-    network: mainnet
-    source:
-      abi: PriceFeed
-      address: "0x4c517D4e2C851CA76d7eC94B805269Df0f2201De"
-      startBlock: 12178551
-    mapping:
-      file: ./src/mapping/PriceFeed.ts
-      language: wasm/assemblyscript
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      entities:
-        - Token
-      abis:
-        - name: PriceFeed
-          file: ./abis/PriceFeed.json
-      eventHandlers:
-        - event: LastGoodPriceUpdated(uint256)
-          handler: handleLastGoodPriceUpdated
-  - kind: ethereum/contract
-    name: ActivePool
-    network: mainnet
-    source:
-      address: "0xDf9Eb223bAFBE5c5271415C75aeCD68C21fE3D7F"
-      abi: ActivePool
-      startBlock: 12178562
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      language: wasm/assemblyscript
-      entities: []
-      abis:
-        - name: ActivePool
-          file: ./abis/ActivePool.json
-      eventHandlers:
-        - event: ActivePoolETHBalanceUpdated(uint256)
-          handler: handleActivePoolETHBalanceUpdated
-        - event: ActivePoolLUSDDebtUpdated(uint256)
-          handler: handleActivePoolLUSDDebtUpdated
-      file: ./src/mapping/ActivePool.ts
-  - kind: ethereum/contract
-    name: CollSurplusPool
-    network: mainnet
-    source:
-      address: "0x3D32e8b97Ed5881324241Cf03b2DA5E2EBcE5521"
-      abi: CollSurplusPool
-      startBlock: 12178578
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      language: wasm/assemblyscript
-      entities: []
-      abis:
-        - name: CollSurplusPool
-          file: ./abis/CollSurplusPool.json
-      eventHandlers:
-        - event: CollBalanceUpdated(indexed address,uint256)
-          handler: handleCollBalanceUpdated
-      file: ./src/mapping/CollSurplusPool.ts
-  - kind: ethereum/contract
-    name: StabilityPool
-    network: mainnet
-    source:
-      address: "0x66017D22b0f8556afDd19FC67041899Eb65a21bb"
-      abi: StabilityPool
-      startBlock: 12178565
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      language: wasm/assemblyscript
-      entities: []
-      abis:
-        - name: StabilityPool
-          file: ./abis/StabilityPool.json
-      eventHandlers:
-        - event: StabilityPoolETHBalanceUpdated(uint256)
-          handler: handleStabilityPoolETHBalanceUpdated
-        - event: StabilityPoolLUSDBalanceUpdated(uint256)
-          handler: handleStabilityPoolLUSDBalanceUpdated
-      file: ./src/mapping/StabilityPool.ts
