--- subgraphs/liquity-ext/src/utils/constants.ts
@@ -1,190 +1,0 @@
-import { BigDecimal, BigInt } from "@graphprotocol/graph-ts";
-
-////////////////////
-///// Versions /////
-////////////////////
-
-export const PROTOCOL_NAME = "Liquity";
-export const PROTOCOL_SLUG = "liquity";
-export const PROTOCOL_SCHEMA_VERSION = "2.0.1";
-export const PROTOCOL_SUBGRAPH_VERSION = "1.2.0";
-export const PROTOCOL_METHODOLOGY_VERSION = "1.0.1";
-
-////////////////////////
-///// Schema Enums /////
-////////////////////////
-
-// The network names corresponding to the Network enum in the schema.
-// They also correspond to the ones in `dataSource.network()` after converting to lower case.
-// See below for a complete list:
-// https://thegraph.com/docs/en/hosted-service/what-is-hosted-service/#supported-networks-on-the-hosted-service
-export namespace Network {
-  export const ARBITRUM_ONE = "ARBITRUM_ONE";
-  export const ARWEAVE_MAINNET = "ARWEAVE_MAINNET";
-  export const AVALANCHE = "AVALANCHE";
-  export const BOBA = "BOBA";
-  export const AURORA = "AURORA";
-  export const BSC = "BSC"; // aka BNB Chain
-  export const CELO = "CELO";
-  export const COSMOS = "COSMOS";
-  export const CRONOS = "CRONOS";
-  export const MAINNET = "MAINNET"; // Ethereum mainnet
-  export const FANTOM = "FANTOM";
-  export const FUSE = "FUSE";
-  export const HARMONY = "HARMONY";
-  export const JUNO = "JUNO";
-  export const MOONBEAM = "MOONBEAM";
-  export const MOONRIVER = "MOONRIVER";
-  export const NEAR_MAINNET = "NEAR_MAINNET";
-  export const OPTIMISM = "OPTIMISM";
-  export const OSMOSIS = "OSMOSIS";
-  export const MATIC = "MATIC"; // aka Polygon
-  export const XDAI = "XDAI"; // aka Gnosis Chain
-}
-
-export namespace ProtocolType {
-  export const EXCHANGE = "EXCHANGE";
-  export const LENDING = "LENDING";
-  export const YIELD = "YIELD";
-  export const BRIDGE = "BRIDGE";
-  export const GENERIC = "GENERIC";
-}
-
-export namespace VaultFeeType {
-  export const MANAGEMENT_FEE = "MANAGEMENT_FEE";
-  export const PERFORMANCE_FEE = "PERFORMANCE_FEE";
-  export const DEPOSIT_FEE = "DEPOSIT_FEE";
-  export const WITHDRAWAL_FEE = "WITHDRAWAL_FEE";
-}
-
-export namespace LiquidityPoolFeeType {
-  export const FIXED_TRADING_FEE = "FIXED_TRADING_FEE";
-  export const TIERED_TRADING_FEE = "TIERED_TRADING_FEE";
-  export const DYNAMIC_TRADING_FEE = "DYNAMIC_TRADING_FEE";
-  export const FIXED_LP_FEE = "FIXED_LP_FEE";
-  export const DYNAMIC_LP_FEE = "DYNAMIC_LP_FEE";
-  export const FIXED_PROTOCOL_FEE = "FIXED_PROTOCOL_FEE";
-  export const DYNAMIC_PROTOCOL_FEE = "DYNAMIC_PROTOCOL_FEE";
-}
-
-export namespace RewardTokenType {
-  export const DEPOSIT = "DEPOSIT";
-  export const BORROW = "BORROW";
-}
-
-export namespace LendingType {
-  export const CDP = "CDP";
-  export const POOLED = "POOLED";
-}
-
-export namespace RiskType {
-  export const GLOBAL = "GLOBAL";
-  export const ISOLATED = "ISOLATED";
-}
-
-export namespace InterestRateType {
-  export const STABLE = "STABLE";
-  export const VARIABLE = "VARIABLE";
-  export const FIXED = "FIXED";
-}
-
-export namespace InterestRateSide {
-  export const LENDER = "LENDER";
-  export const BORROWER = "BORROWER";
-}
-
-export namespace PositionSide {
-  export const LENDER = "LENDER";
-  export const BORROWER = "BORROWER";
-}
-
-//////////////////////////////
-///// Ethereum Addresses /////
-//////////////////////////////
-
-export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
-export const ETH_ADDRESS = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
-
-export const UNISWAP_V2_FACTORY = "0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f";
-
-export const WETH_ADDRESS = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2";
-export const USDC_WETH_PAIR = "0xb4e16d0168e52d35cacd2c6185b44281ec28c9dc"; // created 10008355
-export const DAI_WETH_PAIR = "0xa478c2975ab1ea89e8196811f51a7b7ade33eb11"; // created block 10042267
-export const USDT_WETH_PAIR = "0x0d4a11d5eeaac28ec3f61d100daf4d40471f1852"; // created block 10093341
-
-////////////////////////
-///// Type Helpers /////
-////////////////////////
-
-export const DEFAULT_DECIMALS = 18;
-
-export const USDC_DECIMALS = 6;
-export const USDC_DENOMINATOR = BigDecimal.fromString("1000000");
-
-export const BIGINT_ZERO = BigInt.fromI32(0);
-export const BIGINT_ONE = BigInt.fromI32(1);
-export const BIGINT_TWO = BigInt.fromI32(2);
-export const BIGINT_HUNDRED = BigInt.fromI32(100);
-export const BIGINT_THOUSAND = BigInt.fromI32(1000);
-export const BIGINT_MAX = BigInt.fromString(
-  "115792089237316195423570985008687907853269984665640564039457584007913129639935"
-);
-
-export const INT_NEGATIVE_ONE = -1 as i32;
-export const INT_ZERO = 0 as i32;
-export const INT_ONE = 1 as i32;
-export const INT_TWO = 2 as i32;
-export const INT_FOUR = 4 as i32;
-
-export const BIGDECIMAL_ZERO = new BigDecimal(BIGINT_ZERO);
-export const BIGDECIMAL_ONE = new BigDecimal(BIGINT_ONE);
-export const BIGDECIMAL_TWO = new BigDecimal(BIGINT_TWO);
-export const BIGDECIMAL_HUNDRED = new BigDecimal(BIGINT_HUNDRED);
-
-export const MAX_UINT = BigInt.fromI32(2).times(BigInt.fromI32(255));
-
-/////////////////////
-///// Date/Time /////
-/////////////////////
-
-export const SECONDS_PER_DAY = 60 * 60 * 24; // 86400
-export const SECONDS_PER_HOUR = 60 * 60; // 3600
-export const MS_PER_DAY = new BigDecimal(BigInt.fromI32(24 * 60 * 60 * 1000));
-export const DAYS_PER_YEAR = new BigDecimal(BigInt.fromI32(365));
-export const MS_PER_YEAR = DAYS_PER_YEAR.times(
-  new BigDecimal(BigInt.fromI32(24 * 60 * 60 * 1000))
-);
-
-////////////////
-///// Misc /////
-////////////////
-
-export const ETH_SYMBOL = "ETH";
-export const ETH_NAME = "Ether";
-
-/////////////////////////////
-///// Protocol Specific /////
-/////////////////////////////
-
-export const TROVE_MANAGER = "0xA39739EF8b0231DbFA0DcdA07d7e29faAbCf4bb2";
-
-export const ACTIVE_POOL_CREATED_TIMESTAMP = BigInt.fromI32(1617600130);
-
-export const ACTIVE_POOL_CREATED_BLOCK = BigInt.fromI32(12178562);
-
-export const ACTIVE_POOL = "0xDf9Eb223bAFBE5c5271415C75aeCD68C21fE3D7F";
-
-export const STABILITY_POOL = "0x66017D22b0f8556afDd19FC67041899Eb65a21bb";
-
-export const LUSD_ADDRESS = "0x5f98805A4E8be255a32880FDeC7F6728C6568bA0";
-
-export const MINIMUM_COLLATERAL_RATIO = BigDecimal.fromString("1.1");
-
-export const MAXIMUM_LTV = BIGDECIMAL_ONE.div(MINIMUM_COLLATERAL_RATIO).times(
-  BIGDECIMAL_HUNDRED
-);
-
-export const LIQUIDATION_FEE_PERCENT = BigDecimal.fromString("0.5");
-export const LIQUIDATION_FEE = LIQUIDATION_FEE_PERCENT.div(BIGDECIMAL_HUNDRED);
-
-export const LIQUIDATION_RESERVE_LUSD = BigDecimal.fromString("200");
