--- subgraphs/liquity-ext/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "@graphprotocol/graph-ts/types/tsconfig.base.json",
-  "include": ["src"]
-}
