--- subgraphs/liquity/protocols/liquity/config/networks/ethereum/ethereum.json
@@ -1,1 +1,0 @@
-{}
