+++ subgraphs/liquity/src/entities/position.ts
@@ -6,7 +6,7 @@ import {
   PositionSnapshot,
   _Trove,
 } from "../../generated/schema";
-import { BIGINT_ZERO, INT_ZERO } from "../../src/utils/constants";
+import { BIGINT_ZERO, INT_ZERO } from "../utils/constants";
 import { PositionSide } from "../utils/constants";
 import { getOrCreateAccount } from "./account";
 import {
